# Refactor API frame helpers to enable buffer reuse