https://github.com/esphome/esphome/pull/8805 will heavily conflict. This one has more value so it should merge first.

# What does this implement/fix?


This PR refactors the ESPHome API component's frame helper system to enable efficient buffer reuse by passing `ProtoWriteBuffer` objects instead of raw data pointers. This optimization significantly reduces memory allocations and improves performance for high-frequency message exchanges.

### API Frame Helper Interface Updates

1. **Added frame padding and footer size methods**:
   - `frame_header_padding()`: Returns bytes reserved for protocol headers (7 for Noise, 6 for Plaintext)
   - `frame_footer_size()`: Returns footer size (16 bytes for Noise MAC, 0 for Plaintext)

2. **Renamed and updated write method**:
   - `write_packet()` → `write_protobuf_packet()`
   - Now accepts `ProtoWriteBuffer*` instead of raw data pointers
   - Enables in-place header writing without data copying

### Buffer Management Strategy

The new implementation uses a single reusable buffer with pre-allocated padding:

1. **Pre-allocation**: A single buffer is allocated once and reused for multiple messages
2. **Header Padding**: Buffer starts with pre-inserted padding bytes (7 for Noise, 6 for Plaintext)
3. **Message Encoding**: Protocol buffer data is encoded starting after the padding area
4. **In-place Header Writing**: Headers are written into the pre-allocated padding space without moving data
5. **Footer Space**: Additional space is reserved for protocol footers (Noise MAC)

### Noise performance changes

Before: ~66% performance penalty for encrypted connections --- it was all suffering from memory copy
After: ~7% performance penalty for encrypted connections

base branch `add_api_stats`
testing branch `reuse_buffer_api_stats`
Read is also somehow faster without the other upcoming optmizations but its likely because its not churning memory anymore -- avg read_packet: 0.19ms → 0.07ms (63% faster in new version)

#### Single Loop Period Stats (Last 60s)
- **Old Version**: count=6754, avg=0.50ms, max=204ms, total=3407ms
- **New Version**: count=6862, avg=0.31ms, max=184ms, total=2121ms
- **Improvement**: 
  - Average: 38% faster (0.50ms → 0.31ms)
  - Total time: 37% reduction (3407ms → 2121ms)
  - Max time: 10% reduction (204ms → 184ms)

#### Loop Total Stats (Since Boot)
- **Old Version**: count=256929, avg=0.52ms, max=205ms, total=132979ms
- **New Version**: count=20568, avg=0.31ms, max=184ms, total=6446ms
- **Improvement**: 
  - Average: 40% faster (0.52ms → 0.31ms)
  - Max time: 10% reduction (205ms → 184ms)

### Plaintext performance changes

As expected its not much difference: write_packet operations (6% faster). The overall loop time decreased by ~3.6%

Seems to help a little bit with free memory as well
<img width="1099" alt="Screenshot 2025-05-18 at 12 04 42 AM" src="https://github.com/user-attachments/assets/9e67ed48-a985-4272-ab99-6e7d2e69b64c" />




## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- fixes esphome/issues#6441

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- esphome/esphome-docs#<esphome-docs PR number goes here>

### Testing


```
external_components:
  - source: github://pr#8825
    components: [ api ]
    refresh: 0s
```


## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [x] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# Example config.yaml

```

## Checklist:
  - [ ] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).