@bdraco 2025-05-18
Thanks. I'll fix the conflicts on https://github.com/esphome/esphome/pull/8805 after the PyCon dinner